// lib/enums.ts
export enum TOOLBAR_BTNS {
  THUMBNAILS = 1,
  TYPE = 2,
  NONE = 3,
  DRAW = 4,
  SIGN = 5,
  COMMENT = 6,
  ERASE = 7,
  HIGHLIGHT = 8,
  UNDERLINE = 9,
}

export enum LEFT_SIDEBAR_ENUMS {
  POPULAR = 0,
  ANOTATE = 1,
  UNDERLINE = 3,
  HIGHLIGHT = 4,
}

export enum DOCUMENT_UPLOAD_STATUS {
  PRE_UPLOAD = 1,
  UPLOADING = 2,
  SUCCESS = 3,
  ERROR = 4,
  LOADED_IN_EDITOR = 5,
}